move data feeds api to seperate controller
