+++ api/app/controllers/spree/api/v2/storefront/stores_controller.rb
@@ -7,19 +7,7 @@ module Spree
             render_serialized_payload { serialize_resource(current_store) }
           end
 
-          def data_feeds_google_rss_feed
-            send_data data_feeds_google_rss_service.value[:file], filename: 'products.rss', type: 'text/xml'
-          end
-
           private
-
-          def settings
-            @settings ||= Spree::DataFeedSetting.find_by!(spree_store_id: current_store, uuid: params[:unique_url], enabled: true)
-          end
-
-          def data_feeds_google_rss_service
-            Spree::Dependencies.data_feeds_google_rss_service.constantize.new.call(settings)
-          end
 
           def model_class
             Spree::Store
